Added and customized: - base class for sqlalchemy models; - base repo for database operations; bot.py: - added broadcast handler hello function; - global middleware registration; - session factory and creating tables in the database;
Secure message broadcasting template.

In the bot config I added a database configuration class with environment variables and a method for building an url for the database.

Added bot middleware initiating a session, creating/updating a user for any update and propagating repo and user to data.

tgbot/dialogs/getters.py:
- Added recipe and product classes, as well as methods with template data and then with real data from the database for now only for receipts;

Added orm.py file with AsyncORM class to services to create tables.

Created a model of user, products and a file with class in repo to work with them in the database.

Created a model of receipts and a file with a class in the repo to work with them in the database.

RequestsRepo class in requests.py that combines repo for operations with the database.

tgbot/dialogs/selected.py:
- Added method for processing xml/json files(receipts) and method for selected receipt

- operations of adding a receipt and products to the database

infrastruct/database/setup.py:
- create engine and create session pool methods

Added menu and submenu buttons and transitions: to main, to receipts, back. Display of products from a receipt.